+++ app/src/main/java/woowacourse/movie/domain/movie/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie.model
+package woowacourse.movie.domain.movie
 
 import android.os.Parcelable
 import androidx.annotation.DrawableRes
